feat(config): add config file per environment
The SDK will default to .ropsten as its base environment.

A version file was added internally to enable us read its value and
use it in logs and  other things. The version has been set to 0.1.0
for now.

Note that Swift Package requires all files to be included in the
Source folder.