feat: include StarkCurve params and operations
The curve points and its required calculations to derive from the
StarkCurve has been added.

In order to work with such large numbers the BigInt library has been
included as a dependency.

identifier_name swiftlint rule has been disabled because this package
will contain lots of math operations that are written in a short
variable names as a standard in the equations.