feat: add support to private and public keys
A private key can be created from a BitInt number or a hex, as long
as it's in the range of the StarkCurve's N value.

A public key is derived from the private key by multiplying the
StarkCurve G point by the private key number.

A bunch of extensions were added to ease the work with hex and bytes.