feat: implement stark key generation
The Stark Key pair is generated from a signature and ethereum address.

The signature's s value is used as the seed value while the address is
used as part of a predefined path that's part of the Stark
implementation.

The seed and path are used to derive a private key seed that is then
ground into the final Stark Private Key, which can be used to retrieve
the Stark Public Key.