refactor: make ImmutableXCore instantiable
Structs/Classes with static/class methods are harder to test and
conform to protocols for mocking in Swift.

Making ImmutableXCore being used through its shared singleton
instance (common pattern in iOS) allows for better dependency
injection, as well as, allowing apps to easily extend the struct to
make it conform to a protocol and mock it if desired. For example:

```swift
protocol ImmutableXCoreProtocol {
    func buy(
        orderId: String,
        fees: [FeeEntry],
        signer: Signer,
        starkSigner:
        StarkSigner
    ) async throws -> CreateTradeResponse
}

extension ImmutableXCore: ImmutableXCoreProtocol {}
```