refactor: remove unnecessary methods from Signer
It didnt make sense to have multiple ways of retrieving the same data
since ultimately the SDK only uses the async/await approach. We'd have
to reimplement everything in the old closure style to make use of it,
which is not worth the maintenance. It is very simple for developers
to wrap closure based callbacks in async/await for this particular
scenario.

All other implementations the SDK exposes will continue to support the
old closure style.