docs: remove docs folder
The issue of hosting the static files in our website has been fixed.
The Github Page has been taken down in favour of our website's docs
https://docs.x.immutable.com/sdk-docs/core-sdk-swift/overview

The script to generate the static files has been updated to work with
our website and versioning.

Docs folder and .doccarchive file have been added to gitignore not to
be committed by accident.