feat: update Models templates to always be public
When using the 'nonPublicApi' flag in OpenAPI generator, all clients
and models are made internal by default. However, we wish to have
the clients made internal but the models public, to get the most out
of the generated code, while keeping the interface consistent,
allowing to expose only the method we see fit through ImmutableX
instead of relying on devs using the clients directly.